moved the nodes loader functions
